+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
-- [...]
+- **[UPDATE]** `<Button>` is `disabled` while in loading or checked state
 
 # v0.17.7 (27/11/2018)